gym_trainer_botCopy/pyproject.toml
[project]
name = "gym-trainer-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "Flask",
    "google-generativeai",
    "python-dotenv",
    "Werkzeug",
    "google-genai",
    "psycopg2-binary"
]
